Updates for current version of Sphinx
It wasn't clear from the existing documentation what version of Sphinx it was
built with. Attempting to build the docs with the latest version (1.2.2) with
`make html` caused build errors.

I simply ran `sphinx-quickstart` and let Sphinx rebuild `Makefile` and `conf.py`
itself. I also added an entry to `.gitignore` to ignore the `docs/_build`
directory, which is where Sphinx will now create the built documentation.